Refactoring. Update the social preview image. Update the document nav link to point to docs.zarf
