+++ src/components/PageLayout.tsx
@@ -1,4 +1,4 @@
-import React, { ReactElement, PropsWithChildren, ReactNode } from 'react';
+import React, { ReactElement, PropsWithChildren } from 'react';
 import getPathname from '../hooks/getPathname';
 import ZarfNav from './NavBar/ZarfNav';
 import SiteHelmet from './SiteHelmet';
@@ -6,12 +6,17 @@ import '../assets/css/blobz.min.css';
 import { Box } from '@mui/material';
 import '../assets/css/global.css';
 import Footer from './Footer';
+import { SeoProps } from './seo';
 
-function PageLayout(props: PropsWithChildren<ReactNode>): ReactElement {
+export interface PageLayoutProps {
+  seoProps?: SeoProps;
+}
+
+function PageLayout(props: PropsWithChildren<PageLayoutProps>): ReactElement {
   const pathname = getPathname();
   return (
     <>
-      <SiteHelmet />
+      <SiteHelmet {...props.seoProps} route={pathname.replace('/', '')} />
       <ZarfNav pathname={pathname} />
       <Box
         component="main"
