+++ src/components/Sections/Examples.tsx
@@ -1,0 +1,17 @@
+import { exampleCards } from '../../assets/data/cardData';
+import React, { ReactElement } from 'react';
+import ExampleCard from '../ExampleCard';
+
+function Examples(): ReactElement {
+  return (
+    <>
+      {exampleCards.map(
+        (c, i): ReactElement => (
+          <ExampleCard {...c} key={i} />
+        ),
+      )}
+    </>
+  );
+}
+
+export default Examples;
