+++ src/components/Sections/OpenSourceStats/Stat.tsx
@@ -1,5 +1,5 @@
-import { Box, Typography } from '@mui/material';
 import React, { MutableRefObject, ReactElement } from 'react';
+import { Box, Typography } from '@mui/material';
 
 interface StatProps {
   title: string | number;
